Extend testsuite to support better block issuing with slots
